requirements.txt
langchain_chroma==1.1.0
langchain_classic==1.0.0
langchain_core==1.2.1
langchain_groq==1.1.1
langchain_huggingface==1.2.0
python-dotenv==1.2.1
streamlit==1.52.1
